# server/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Photo Diary API",
    "description": "Self-hosted photo-diary backend. Schema is generated from TypeBox-validated route handlers.",
    "version": "0.16.0"
  },
  "components": {
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "scheme": "bearer",
        "description": "JWT issued by POST /api/v1/tokens. Send as `Authorization: Bearer <token>`."
      }
    },
...
  },
  "paths": {
...
    "/api/v1/galleries": {
      "get": {
        "summary": "List galleries visible to the requester",
        "tags": [
          "galleries"
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "required": [
                      "id",
                      "hideMap"
                    ],
                    "properties": {
                      "id": {
                        "type": "string"
                      },
                      "hideMap": {
                        "type": "boolean"
                      }
                    },
                    "additionalProperties": true
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a gallery (admin)",
        "tags": [
          "galleries"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "id"
                ],
                "properties": {
                  "id": {
                    "type": "string",
                    "minLength": 1,
                    "pattern": "^[a-z0-9][a-z0-9_-]*$"
                  },
                  "title": {
                    "type": "string"
                  },
                  "description": {
                    "type": "string"
                  },
                  "titleLocalized": {
                    "type": "object",
                    "additionalProperties": {
                      "type": "string"
                    }
                  },
                  "descriptionLocalized": {
                    "type": "object",
                    "additionalProperties": {
                      "type": "string"
                    }
                  },
                  "defaultLanguage": {
                    "type": "string",
                    "minLength": 1
                  },
                  "icon": {
                    "type": "string"
                  },
                  "epoch": {
                    "type": "string"
                  },
                  "epochType": {
                    "type": "string",
                    "enum": [
                      "birthday",
                      "1-index",
                      "0-index"
                    ]
                  },
                  "theme": {
                    "type": "string",
                    "enum": [
                      "blue",
                      "red",
                      "grayscale",
                      "contrast",
                      "alert",
                      "dark",
                      "amoled",
                      "forest",
                      "silver",
                      "showcase",
                      "teal",
                      "paper",
                      "amber",
                      "lavender",
                      "sage",
                      "slate",
                      "midnight",
                      "espresso"
                    ]
                  },
                  "initialView": {
                    "type": "string",
                    "enum": [
                      "year",
                      "month",
                      "day",
                      "photo"
                    ]
                  },
                  "hostname": {
                    "type": "string"
                  },
                  "sources": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "minLength": 1
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
...
    "/api/v1/galleries/{galleryId}/stats": {
      "post": {
        "summary": "Aggregated stats for a gallery (gallery view)",
        "tags": [
          "stats"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "filter": {
                    "type": "object",
                    "additionalProperties": {
                      "type": "object",
                      "additionalProperties": {
                        "type": "array",
                        "items": {
                          "anyOf": [
                            {
                              "type": "string"
                            },
                            {
                              "type": "number"
                            },
                            {
                              "type": "boolean"
                            },
                            {
                              "type": "null"
                            }
                          ]
                        }
                      }
                    }
                  },
                  "dateRange": {
                    "type": "object",
                    "properties": {
                      "from": {
                        "type": "string",
                        "pattern": "^\\d{4}-\\d{2}-\\d{2}$"
                      },
                      "to": {
                        "type": "string",
                        "pattern": "^\\d{4}-\\d{2}-\\d{2}$"
                      }
                    },
                    "additionalProperties": false
                  },
                  "numericRanges": {
                    "type": "object",
                    "additionalProperties": {
                      "type": "object",
                      "properties": {
                        "min": {
                          "type": "number"
                        },
                        "max": {
                          "type": "number"
                        }
                      },
                      "additionalProperties": false
                    }
                  },
                  "lang": {
                    "type": "string",
                    "minLength": 2,
                    "maxLength": 8
                  }
                }
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "galleryId",
            "required": true
          }
        ],
        "security": [
          {
            "bearer": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "total",
                    "geotaggedCount",
                    "byCategory",
                    "byYearMonth",
                    "summary",
                    "daysInYear",
                    "daysInYearMonth",
                    "byStateCountry",
                    "byCityCountry",
                    "byCityLocalized",
                    "categoryValues",
                    "byGallery"
                  ],
                  "properties": {
                    "total": {
                      "type": "number"
                    },
                    "geotaggedCount": {
                      "type": "number"
                    },
                    "byCategory": {
                      "type": "object",
                      "additionalProperties": {
                        "type": "object",
                        "additionalProperties": {
                          "type": "number"
                        }
                      }
                    },
                    "byYearMonth": {
                      "type": "object",
                      "additionalProperties": {
                        "type": "object",
                        "additionalProperties": {
                          "type": "number"
                        }
                      }
                    },
                    "summary": {
                      "type": "object",
                      "required": [
                        "spanDays",
                        "spanYears",
                        "spanMonths",
                        "peakShape",
                        "variety"
                      ],
                      "properties": {
                        "first": {
                          "type": "string"
                        },
                        "last": {
                          "type": "string"
                        },
                        "spanDays": {
                          "type": "number"
                        },
                        "spanYears": {
                          "type": "number"
                        },
                        "spanMonths": {
                          "type": "number"
                        },
                        "peakShape": {
                          "type": "object",
                          "additionalProperties": {
                            "type": "string"
                          }
                        },
                        "variety": {
                          "type": "object",
                          "additionalProperties": {
                            "type": "number"
                          }
                        }
                      }
                    },
                    "daysInYear": {
                      "type": "object",
                      "additionalProperties": {
                        "type": "number"
                      }
                    },
                    "daysInYearMonth": {
                      "type": "object",
                      "additionalProperties": {
                        "type": "object",
                        "additionalProperties": {
                          "type": "number"
                        }
                      }
                    },
                    "byStateCountry": {
                      "type": "object",
                      "additionalProperties": {
                        "type": "string"
                      }
                    },
                    "byCityCountry": {
                      "type": "object",
                      "additionalProperties": {
                        "type": "string"
                      }
                    },
                    "byCityLocalized": {
                      "type": "object",
                      "additionalProperties": {
                        "type": "string"
                      }
                    },
                    "categoryValues": {
                      "type": "object",
                      "additionalProperties": {
                        "type": "array",
                        "items": {
                          "type": "string"
                        }
                      }
                    },
                    "byGallery": {
                      "type": "object",
                      "additionalProperties": {
                        "type": "number"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/galleries/{galleryId}/stats/evolution": {
      "post": {
        "summary": "Per-bucket time-series for a trendable category",
        "tags": [
          "stats"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "category"
                ],
                "properties": {
                  "category": {
                    "type": "string",
                    "minLength": 1
                  },
                  "filter": {
                    "type": "object",
                    "additionalProperties": {
                      "type": "object",
                      "additionalProperties": {
                        "type": "array",
                        "items": {
                          "anyOf": [
                            {
                              "type": "string"
                            },
                            {
                              "type": "number"
                            },
                            {
                              "type": "boolean"
                            },
                            {
                              "type": "null"
                            }
                          ]
                        }
                      }
                    }
                  },
                  "dateRange": {
                    "type": "object",
                    "properties": {
                      "from": {
                        "type": "string",
                        "pattern": "^\\d{4}-\\d{2}-\\d{2}$"
                      },
                      "to": {
                        "type": "string",
                        "pattern": "^\\d{4}-\\d{2}-\\d{2}$"
                      }
                    },
                    "additionalProperties": false
                  },
                  "numericRanges": {
                    "type": "object",
                    "additionalProperties": {
                      "type": "object",
                      "properties": {
                        "min": {
                          "type": "number"
                        },
                        "max": {
                          "type": "number"
                        }
                      },
                      "additionalProperties": false
                    }
                  },
                  "lang": {
                    "type": "string",
                    "minLength": 2,
                    "maxLength": 8
                  }
                }
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "galleryId",
            "required": true
          }
        ],
        "security": [
          {
            "bearer": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "yearMonths",
                    "buckets"
                  ],
                  "properties": {
                    "yearMonths": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    },
                    "buckets": {
                      "type": "object",
                      "additionalProperties": {
                        "type": "object",
                        "required": [
                          "counts",
                          "cumulative"
                        ],
                        "properties": {
                          "counts": {
                            "type": "array",
                            "items": {
                              "type": "number"
                            }
                          },
                          "cumulative": {
                            "type": "array",
                            "items": {
                              "type": "number"
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/stats": {
      "post": {
        "summary": "Aggregated stats across all galleries (admin)",
        "tags": [
          "stats"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "filter": {
                    "type": "object",
                    "additionalProperties": {
                      "type": "object",
                      "additionalProperties": {
                        "type": "array",
                        "items": {
                          "anyOf": [
                            {
                              "type": "string"
                            },
                            {
                              "type": "number"
                            },
                            {
                              "type": "boolean"
                            },
                            {
                              "type": "null"
                            }
                          ]
                        }
                      }
                    }
                  },
                  "dateRange": {
                    "type": "object",
                    "properties": {
                      "from": {
                        "type": "string",
                        "pattern": "^\\d{4}-\\d{2}-\\d{2}$"
                      },
                      "to": {
                        "type": "string",
                        "pattern": "^\\d{4}-\\d{2}-\\d{2}$"
                      }
                    },
                    "additionalProperties": false
                  },
                  "numericRanges": {
                    "type": "object",
                    "additionalProperties": {
                      "type": "object",
                      "properties": {
                        "min": {
                          "type": "number"
                        },
                        "max": {
                          "type": "number"
                        }
                      },
                      "additionalProperties": false
                    }
                  },
                  "lang": {
                    "type": "string",
                    "minLength": 2,
                    "maxLength": 8
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "total",
                    "geotaggedCount",
                    "byCategory",
                    "byYearMonth",
                    "summary",
                    "daysInYear",
                    "daysInYearMonth",
                    "byStateCountry",
                    "byCityCountry",
                    "byCityLocalized",
                    "categoryValues",
                    "byGallery"
                  ],
                  "properties": {
                    "total": {
                      "type": "number"
                    },
                    "geotaggedCount": {
                      "type": "number"
                    },
                    "byCategory": {
                      "type": "object",
                      "additionalProperties": {
                        "type": "object",
                        "additionalProperties": {
                          "type": "number"
                        }
                      }
                    },
                    "byYearMonth": {
                      "type": "object",
                      "additionalProperties": {
                        "type": "object",
                        "additionalProperties": {
                          "type": "number"
                        }
                      }
                    },
                    "summary": {
                      "type": "object",
                      "required": [
                        "spanDays",
                        "spanYears",
                        "spanMonths",
                        "peakShape",
                        "variety"
                      ],
                      "properties": {
                        "first": {
                          "type": "string"
                        },
                        "last": {
                          "type": "string"
                        },
                        "spanDays": {
                          "type": "number"
                        },
                        "spanYears": {
                          "type": "number"
                        },
                        "spanMonths": {
                          "type": "number"
                        },
                        "peakShape": {
                          "type": "object",
                          "additionalProperties": {
                            "type": "string"
                          }
                        },
                        "variety": {
                          "type": "object",
                          "additionalProperties": {
                            "type": "number"
                          }
                        }
                      }
                    },
                    "daysInYear": {
                      "type": "object",
                      "additionalProperties": {
                        "type": "number"
                      }
                    },
                    "daysInYearMonth": {
                      "type": "object",
                      "additionalProperties": {
                        "type": "object",
                        "additionalProperties": {
                          "type": "number"
                        }
                      }
                    },
                    "byStateCountry": {
                      "type": "object",
                      "additionalProperties": {
                        "type": "string"
                      }
                    },
                    "byCityCountry": {
                      "type": "object",
                      "additionalProperties": {
                        "type": "string"
                      }
                    },
                    "byCityLocalized": {
                      "type": "object",
                      "additionalProperties": {
                        "type": "string"
                      }
                    },
                    "categoryValues": {
                      "type": "object",
                      "additionalProperties": {
                        "type": "array",
                        "items": {
                          "type": "string"
                        }
                      }
                    },
                    "byGallery": {
                      "type": "object",
                      "additionalProperties": {
                        "type": "number"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}